// src/domain/enums/GameType.ts
export enum GameType {
  AttackAndDefense = 'Attack & Defense',
  ClanArena = 'Clan Arena',
  CaptureTheFlag = 'Capture The Flag',
  Duel = 'Duel',
  Domination = 'Domination',
  FreeForAll = 'Free For All',
  FreezeTag = 'Freeze Tag',
  Harvester = 'Harvester',
  OneFlagCtf = 'One-Flag CTF',
  Race = 'Race',
  RedRover = 'Red Rover',
  TeamDeathmatch = 'Team Deathmatch'
}